Bugs fixed, jackpot enabled, whole bot runs on decimals now.

+++ Basebot/src/org/smokinmils/bot/Bet.java
@@ -5,12 +5,12 @@ import org.smokinmils.database.types.ProfileType;
 public class Bet {
 
 	private String user;
-	private int amount;
+	private double amount;
 	private String choice;
 	private boolean valid;
 	private ProfileType profile;
 	
-	public Bet(String user, ProfileType profile, int amount, String choice) {
+	public Bet(String user, ProfileType profile, double amount, String choice) {
 		this.user = user;
 		this.amount = amount;
 		this.choice = choice;
@@ -27,7 +27,7 @@ public class Bet {
 	}
 	
 	public String getUser() { return user; }
-	public int getAmount()	{ return amount; }
+	public double getAmount()	{ return amount; }
 	public String getChoice() { return choice; }
 	public boolean isValid() { return valid; }
 	public ProfileType getProfile() { return profile; }
